Changes to LocationSchema
Changes Made to

user.js:

Made some changes in the definiton of the LocationSchema so that it doesn't create an _id attribute for our geoJSON objects, also forcing a spatial index on the Location attribute so that we'll be able to take advantage of spatial queries on our User data.

app.js:

Changed the ordering of the attributes obtained in POST, which seems to insert document attributes in the correct order now (aside from location, in which coordinates still appears first, then type).  Shouldn't be that big of deal but would be nice to have them in order.