Cleaning up recent changes
Deleted commented out code in our app.js file, standardardized the different
routes in our routes path to use router.post or router.get rather than using router.route also added the router.use middleware function just incase it's needed, not sure it is but most examples out there make use of it, in particular the next() function inside of it.  Moved location.js from /models to /models/schema to better reflect that it exports a schema and not a model.  Also added the dateFormat package so that our dates are formatted nicely when outputting to console.

 Changes to be committed:
	modified:   app.js
	modified:   index.js
	renamed:    models/location.js -> models/schemas/location.js
	modified:   models/studentlocation.js
	modified:   models/tutorlocation.js
	modified:   models/user.js
	modified:   package.json
	deleted:    routes/index.js
	modified:   routes/studentlocations.js
	modified:   routes/students.js
	modified:   routes/tutorlocations.js
	modified:   routes/tutors.js
	modified:   routes/users.js