Added New Route to Export to KML
New Route added that exports user locations to KML
Useful for viewing user locations on map in Google Earth.

Also changed the coordinates of the FindUserTest so that
their coordinates are within the Continental US. Having
coordinates so far away causes Google Map to position
default extent for layer between the two populations of points.
Tests have been updated to reflect this change.

 Changes to be committed:
	modified:   package.json
	modified:   routes/users.js
	modified:   test/users.js