Updated booking model, user model and booking and user routes
Added student_name and tutor_name to booking model, changed user fcm_token property to be fcm_tokens property, which is now a collection to store multiple fcm tokens, this will allow users to login on multiple devices and receive messages.

Bluebird was also added as a dependency, not using any particular functionality from it yet, but may use its implementation of promises.

  Changes to be committed:
	modified:   models/booking.js
	modified:   models/user.js
	modified:   package.json
	modified:   routes/booking.js
	modified:   routes/users.js