Continued Integration with timekit.io
Made changes to the user and booking model to accomedate timekit.io. User model now has a timekit_calendar_id which identifies a user's own unique calendar for bookings on the timekit.io side.  The Booking model (which may end up being redundant, we need to discuss further whether to keep it or get rid of it) now has a property for scheduledendingdate to record the date/time in ISO format of the time the meeting/booking ends, it also has an id for the booking on the timekit.io side called timekit_booking_id.

user have been further tweeked there are now both GET and POST routes for /me/timekit and /me/timekit/calendar.

On a GET /me/timekit returns timekit information for current user logged in via stormpath otherwise it returns status code 500. On a POST /me/timekit a new timekit.io user and profile is created for the current user logged in via stormpath, assigning the user the their own personal timekit io token needed to access the timekit.io api, this doesn't appear to expire.

On GET /me/timekit/calendar a user's calendar is displayed if a timekit_calendar_id exists for their mongo document, if not status code 500 is returned.  On POST /me/timekit/calendar a user that has a timekit token assigned has a new calendar generated for them.  For now we're only allowing one timekit calendar to be created per Two2er user, so if a calendar already exists a status code error 500 is returned (May tweak the status code in refactoring to better indicate action).

On POST to /booking/request a timekit.io event is generated for a user in their timekit.io calendar if it exists, some information is canned at the moment, we will need to work on adding fields perhaps on the UI side to accomedate the information timekit.io wants here.

  Changes to be committed:
	modified:   models/booking.js
	modified:   models/studentlocation.js
	modified:   models/user.js
	modified:   routes/booking.js
	modified:   routes/users.js