Stormpath Client getMyToken Route Added
Added new functionality to the stormpath client that allows us to get StormPath users tokens without having to change the username and password in the stormpathclient.js.  Instead now when loading up our API we can visit the endpoint /dev/getMyToken when you are presented with a form for entering in a username and password.  On submit your token is returned if such a user exists and the password is correct, otherwise an error is returned.

 Changes to be committed:
	modified:   routes/dev.js
	modified:   stormpathclient.js
	new file:   views/token.pug