Updated User Routes
Added logic to the POST on users route so that timekit_token and timekit_calendar_id are generated
when a new user document is created.  Also a new "blank" student or tutor embedded document is created
upon the creation of a new user document based on the userMode attribute sent on POST.  All new users documents now have a corresponding timekit.io account and default calendar.

Removed getUserByName route

Isn't being used in our application anywhere, in addition it's possible for full names to not be unique, so find here isn't particularly useful for finding a unique user.

Changed deleteByEmail and deleteById Routes from HTTP GET to HTTP DELETE

We shouldn't be allowing delete's in the database just by visiting a URL in a browser, these should be
requested by DELETE and not GET.