Subject Model and logic added
Added new Mongoose Subject Model to represent a unique subject that a student can find tutoring for.  Also added logic so that when a tutor lists a subject they teach in their profile a corresponding entry is added to the Subjects database collection.

Also added a helper class called TutorSubjects that parses a tutor's collection of subjects represented as strings into Subjects as represented in the Subject model.  Helper class takes this parsed information and then bulk inserts it into our Subject collection.  Subjects that already exist in the collection are ignored as each name attribute in the Subject model must be unique.

Added routes for api/subjects and apiauth/subjects, right now only a GET operation is permitted to display list of subjects in our MongoDB Subjects collection.

Updated pug views to show new routes.

  Changes to be committed:
	modified:   app.js
	new file:   helpers/tutorsubjects.js
	new file:   models/subject.js
	modified:   routes/index.js
	new file:   routes/subjects.js
	modified:   routes/tutors.js
	modified:   routes/users.js
	modified:   views/index.pug