Changes to booking endpoints and re-exposing api operation listing via /api
Booking routes now has two different endpoints for the various Timekit.io
booking actions on an existing bookin. These actions only allow for the state to change on bookings assigned to a specific tutor, if this tutor does not have a timekit_token_id assigned these actions will not work.  Furthermore a student cannot call these endpoints and expect their actions to be applied to the booking, even is it is a booking that includes them.  The different endpoints vary in route based on whether they're called by timekit.io's webhooks or by our users in stormpath.  Routes that are prefixed with booking/timekit/<state> are for timekit, routes that are not prefixed with timekit are meant to be called by a tutor that is authenticated against stormpath.

Api listing of operations is now available again via <root>/api (/apiauth does not display this we can change this if need be)

  Changes to be committed:
	modified:   app.js
	modified:   routes/booking.js