Modified Users Route to control creation of timekit users
After further clarification it was discovered that timekit costs more than we had anticipated as such we are limiting the amount of timekit enabled users to be 12, also all 12 timekit users will be tutors, rather than students.

User endpoints have been updated so that in order to create a timekit user the property enableTimekit must be sent with the request and set to true also the userMode must be set to Tutor if a timekit user is to be created otherwise timekit user creation is skipped.

In addition to this, a user's password must be sent to the timekit endpoint so that for management purposes (e.g. confirming/decling/cancelling bookings) as well as refreshing tokens can be done through our API, without this we lose this functionality.  If a password is not sent in the request the default password for all timekit enabled users is 'two2er1234'.

The /users/me/timekit/auth endpoint was added for the purpose of refreshing timekit enabled users' tokens after they have expired.  If a token is about to expire then the token is automatically refreshed and saved in our user model, if not then the current user's token stays in place.  In either case a response is returned with the number of days left before the user's token expires as well as our user's user document.  This endpoint should be called each time after logging into our android front-end.

The TimekitUserExpiry helper class was created for the purpose of decoupling some of the token expiration logic when checking to see if a token needs to be refreshed in the auth endpoint.

 Changes to be committed:
	new file:   helpers/timekituserexpiry.js
	modified:   helpers/tutorsubjects.js
	modified:   models/user.js
	modified:   routes/users.js